fiddling

+++ Client/Client/player.cpp
@@ -21,7 +21,7 @@ void Player::displayPlayer(){
 	static ImGuiTableFlags flags = ImGuiTableFlags_ScrollX | ImGuiTableFlags_ScrollY | ImGuiTableFlags_RowBg | ImGuiTableFlags_BordersOuter | ImGuiTableFlags_BordersV | ImGuiTableFlags_Reorderable | ImGuiTableFlags_Hideable;
 	const ImVec2 outer_size = ImVec2(0.0f, ImGui::GetTextLineHeightWithSpacing() * 7);
 
-	if (ImGui::BeginTable(std::string("##" + name).c_str(), history.size(), flags, outer_size)) {
+	if (ImGui::BeginTable(std::string("##" + name).c_str(), (int)history.size(), flags, outer_size)) {
 
 		//ImGui::SetColumnWidth(-1, 100.0);
